Refresh prereq

+++ ytdl_win_v6.ps1
@@ -30,7 +30,7 @@ function DownloadNow { # define function to download section
     }
     else {
         Write-Host "`nDefaults (YouTube Video) selected. Please specify any of special resolution below.";
-        Write-Host "`nAV1 offers lower file size, but come in cost of high CPU usage.";
+        Write-Host "`nAV1 offers lower file size, but come in cost of high CPU usage (AV1 only available on select videos)";
         Write-Host "`n[2V1] 2560x1440 resolution AV1 codec`n[4V1] 3840x2160 resolution AV1 codec";
         Write-Host "`nVP9 offers bigger file size, but come at a cost of lower CPU usage.";
         Write-Host "`n[2P9] 2560x1440 resolution VP9 codec`n[4P9] 3840x2160 resolution VP9 codec";
@@ -65,9 +65,14 @@ function DownloadNow { # define function to download section
 
 function DonwloadDone () {
     Write-Host "`n`n`nDownload done! Video is available in Desktop\YouTube_Downloader\`n" -ForegroundColor Green;
-    $returnhome= Read-Host -Prompt "[O] Show in Folder     [other_key] Return to Main Menu";
+    $returnhome= Read-Host -Prompt "If the above download process fails/buggy, you can reinstall package by`n[R] Reinstall/update prerequisites`n`nOtherwise:`n[P] Show in Folder`n[others/blank] Return to Main Menu";
     if ($returnhome -eq "O") {
         Invoke-Item "~\Desktop\YouTube_Downloads\";
+        Clear-Host; MainMenu;
+    }
+    elseif ($returnhome -eq "R") {
+        Write-Host "`n`nUpdating prerequisites..." -ForegroundColor Yellow;
+        RefreshPrereq;
         Clear-Host; MainMenu;
     }
     else {
@@ -76,28 +81,33 @@ function DonwloadDone () {
 }
 
 function BeforeDownloadRoutine {
-    Write-Host "`n`nChecking prerequisites..." -ForegroundColor Yellow
+    Write-Host "`n`nChecking prerequisites..." -ForegroundColor Yellow;
     if ((Test-Path -Path "~\ffmpeg-master-latest-win64-gpl-shared\bin\yt-dlp.exe") -eq $True) {
         Write-Host "`nYou already have the file required.`n" -ForegroundColor Green; 
         Set-Location ~\ffmpeg-master-latest-win64-gpl-shared\bin\; 
     }
     else {
         Write-Host "`nYour computer don't have the required file to run.`nAllow program to download it first." -ForegroundColor DarkMagenta;
-	    # Delete remaining prerequisites first to avoid conflicts
-        Remove-Item "~\ffmpeg.zip" -Recurse -Force; Remove-Item "~\ffmpeg-master-latest-win64-gpl-shared\" -Recurse -Force;
-        # Download ffmpeg win64
-        Invoke-WebRequest -Uri https://github.com/BtbN/FFmpeg-Builds/releases/download/latest/ffmpeg-master-latest-win64-gpl-shared.zip -OutFile .\ffmpeg.zip; 
-	    # Extract ffmpeg.zip archive
-        Expand-Archive -LiteralPath "~\ffmpeg.zip" -DestinationPath "~\"; 
-        # Change active directory to extracted
-        Set-Location ~\ffmpeg-master-latest-win64-gpl-shared\bin\;
-        # Now download yt-dlp binary
-        Invoke-WebRequest -Uri https://github.com/yt-dlp/yt-dlp/releases/latest/download/yt-dlp.exe -OutFile .\yt-dlp.exe;
-        # Delete remaining ffmpeg.zip (we dont need it anymore)
-        Remove-Item "~\ffmpeg.zip" -Recurse -Force;
-        # Message downloads prereq done
-        Write-Host "`nPrerequisites downloads done!" -ForegroundColor Green;
+	    RefreshPrereq;
     }
+}
+
+function RefreshPrereq () {
+    # if statements to delete (this to avoid erroring on client side)
+    if ((Test-Path -Path "~\ffmpeg.zip") -eq $True) {Remove-Item "~\ffmpeg.zip" -Recurse -Force;}
+    if ((Test-Path -Path "~\ffmpeg-master-latest-win64-gpl-shared\bin\yt-dlp.exe") -eq $True) {Remove-Item "~\ffmpeg-master-latest-win64-gpl-shared\" -Recurse -Force;}
+    # Download ffmpeg win64
+    Invoke-WebRequest -Uri https://github.com/BtbN/FFmpeg-Builds/releases/download/latest/ffmpeg-master-latest-win64-gpl-shared.zip -OutFile .\ffmpeg.zip; 
+	# Extract ffmpeg.zip archive
+    Expand-Archive -LiteralPath "~\ffmpeg.zip" -DestinationPath "~\"; 
+    # Change active directory to extracted
+    Set-Location ~\ffmpeg-master-latest-win64-gpl-shared\bin\;
+    # Now download yt-dlp binary
+    Invoke-WebRequest -Uri https://github.com/yt-dlp/yt-dlp/releases/latest/download/yt-dlp.exe -OutFile .\yt-dlp.exe;
+    # Delete remaining ffmpeg.zip (we dont need it anymore)
+    Remove-Item "~\ffmpeg.zip" -Recurse -Force;
+    # Message downloads prereq done
+    Write-Host "`nPrerequisites downloads done!" -ForegroundColor Green;
 }
 
 # Program actually starts here
